refactor: refactoring code, split the component into subcomponents
