+++ lib/components/molecules/Pagination/Pagination.tsx
@@ -1,21 +1,14 @@
-import { useState, useEffect, useMemo, useCallback } from 'react';
+
 import clsx from 'clsx';
 import styles from './Pagination.module.scss';
+import { Select } from 'components/molecules/Select-box/Select';
+import { Typography } from 'components/atoms';
+import { PaginationProps } from './Pagination.types';
+
+import { ArrowBackSimple, ArrowForwardSimple } from 'assets/icons';
+import { usePagination } from './hooks/usePagination/usePagination';
 import { PaginationButton } from './PaginationButton/PaginationButton';
 import { PaginationEllipsis } from './PaginationEllipsis/PaginationEllipsis';
-import { Select } from 'components/molecules/Select-box/Select';
-import { ArrowBackSimple, ArrowForwardSimple } from 'assets/icons';
-import { Typography } from 'components/atoms';
-
-interface PaginationProps {
-  currentPage?: number;
-  totalItems?: number;
-  itemsPerPage?: number;
-  onPageChange: (page: number) => void;
-  onItemsPerPageChange?: (count: number) => void;
-  className?: string;
-}
-type PageItem = number | { type: 'ellipsis'; position: 'left' | 'right' };
 
 export const Pagination = ({
   currentPage = 1,
@@ -25,69 +18,24 @@ export const Pagination = ({
   onItemsPerPageChange,
   className,
 }: PaginationProps) => {
-  const totalPages = Math.max(1, Math.ceil(totalItems / itemsPerPage));
-  const [inputPage, setInputPage] = useState(currentPage.toString());
-  const [activeEllipsis, setActiveEllipsis] = useState<'left' | 'right' | null>(null);
-
-  useEffect(() => {
-    setInputPage(currentPage.toString());
-    setActiveEllipsis(null);
-  }, [currentPage]);
-
-  const handlePageInputChange = useCallback((value: string) => {
-    if (/^\d*$/.test(value)) {
-      setInputPage(value);
-    }
-  }, []);
-
-  const handlePageInputBlur = useCallback(() => {
-    const page = Math.max(1, Math.min(parseInt(inputPage) || currentPage, totalPages));
-    onPageChange(page);
-    setInputPage(page.toString());
-    setActiveEllipsis(null);
-  }, [inputPage, currentPage, totalPages, onPageChange]);
-
-  const handleKeyDown = useCallback((e: React.KeyboardEvent) => {
-    if (e.key === 'Enter') handlePageInputBlur();
-  }, [handlePageInputBlur]);
-
-  const handleItemsPerPageChange = useCallback((selectedValue: number) => {
-    if (selectedValue) {
-      console.log('Changing items per page to:', selectedValue);
-      onItemsPerPageChange?.(selectedValue);
-      onPageChange(1);
-    }
-  }, [onItemsPerPageChange, onPageChange]);
-
-  const handleEllipsisClick = useCallback((position: 'left' | 'right') => {
-    setActiveEllipsis(position);
-  }, []);
-
-  const visiblePages = useMemo<PageItem[]>(() => {
-    if (totalPages <= 1) return [1];
-
-    const pages: PageItem[] = [1];
-    const maxVisiblePages = 5;
-    const boundaryPages = 1;
-
-    if (totalPages <= maxVisiblePages) {
-      for (let i = 2; i <= totalPages; i++) pages.push(i);
-    } else {
-      const startPage = Math.max(2, currentPage - boundaryPages);
-      const endPage = Math.min(totalPages - 1, currentPage + boundaryPages);
-
-      if (startPage > 2) pages.push({ type: 'ellipsis', position: 'left' });
-      for (let i = startPage; i <= endPage; i++) pages.push(i);
-      if (endPage < totalPages - 1) pages.push({ type: 'ellipsis', position: 'right' });
-
-      pages.push(totalPages);
-    }
-
-    return pages;
-  }, [totalPages, currentPage]);
-
-  const pageSizeOptions = useMemo(() => [10, 20, 30, 50, 100], []);
-
+  const {
+    inputPage,
+    activeEllipsis,
+    totalPages,
+    visiblePages,
+    selectOptions,
+    handlePageInputChange,
+    handlePageInputBlur,
+    handleKeyDown,
+    handleItemsPerPageChange,
+    handleEllipsisClick,
+  } = usePagination({
+    currentPage,
+    totalItems,
+    itemsPerPage,
+    onPageChange,
+    onItemsPerPageChange,
+  });
 
   return (
     <div className={clsx(styles.paginationRoot, className)}>
@@ -149,10 +97,7 @@ export const Pagination = ({
             </Typography>
 
             <Select
-              items={pageSizeOptions.map(option => ({
-                value: option.toString(),
-                label: option.toString()
-              }))}
+              items={selectOptions}
               value={itemsPerPage.toString()}
               onValueChange={(value) => {
                 if (value) handleItemsPerPageChange(Number(value));
@@ -178,7 +123,3 @@ export const Pagination = ({
     </div>
   );
 };
-
-
-
-
