+++ lib/components/molecules/Pagination/Pagination.types.ts
@@ -1,8 +1,10 @@
+export type PageItem = number | { type: 'ellipsis'; position: 'left' | 'right' };
+
 export interface PaginationProps {
-  currentPage?: number
-  totalItems?: number
-  itemsPerPage?: number
-  onPageChange: (page: number) => void
-  onItemsPerPageChange?: (count: number) => void
-  className?: string
+  currentPage?: number;
+  totalItems?: number;
+  itemsPerPage?: number;
+  onPageChange: (page: number) => void;
+  onItemsPerPageChange?: (count: number) => void;
+  className?: string;
 }
