+++ lib/components/molecules/Pagination/constants/constants.tsx
@@ -1,0 +1,1 @@
+export const PAGE_SIZE_OPTIONS = [10, 20, 30, 50, 100];
